除錯 - 統一金流訂單回應時加入訂閱外掛類別判斷

+++ includes/payuni/src/gateways/Subscription.php
@@ -44,12 +44,12 @@ class Subscription {
 	/**
 	 * Process subscription payment
 	 *
-	 * @param float    $amount The amount.
+	 * @param int      $amount The amount.
 	 * @param WC_Order $order  The order object.
 	 *
 	 * @return void
 	 */
-	public function process_subscription_payment( float $amount, WC_Order $order ): void {
+	public function process_subscription_payment( int $amount, WC_Order $order ): void {
 		$request = new Request( new CreditSubscription() );
 		$request->build_subscription_request( $amount, $order );
 	}
